+++ src/muncher/items.js
@@ -68,7 +68,7 @@ async function generateImportItems(items) {
     }
   };
   let itemSpells = []; // here we need to parse each available spell and build a mock spell parser
-  const inventory = getInventory(mockDDB, mockCharacter, itemSpells);
+  const inventory = await getInventory(mockDDB, mockCharacter, itemSpells);
   const results = {
     items: inventory,
     itemSpellNames: itemSpells, // this needs to be a list of spells to find